day4 finished

+++ src/main/java/com/group/libraryapp/controller/user/UserController.java
@@ -41,6 +41,12 @@ public class UserController {
 
     @PutMapping("/user")
     public void updateUser(@RequestBody UserUpdateRequest request){
+        String readSql = "SELECT * FROM user WHERE id = ?";
+        boolean isUserNotExist = jdbcTemplate.query(readSql, (rs,rowNum)-> 0, request.getId()).isEmpty();
+        if (isUserNotExist){
+            throw new IllegalArgumentException();
+        }
+
         String sql = "UPDATE user SET name = ? WHERE id = ?";
         jdbcTemplate.update(sql, request.getName(),request.getId());
 
@@ -48,7 +54,14 @@ public class UserController {
 
     @DeleteMapping("/user")
     public void deleteUser(@RequestParam String name){
+        String readSql = "SELECT * FROM user WHERE name = ?";
+        boolean isUserNotExist = jdbcTemplate.query(readSql, (rs,rowNum)-> 0, name).isEmpty();
+        if (isUserNotExist){
+            throw new IllegalArgumentException();
+        }
         String sql = "DELETE from user WHERE name = ?";
         jdbcTemplate.update(sql,name);
     }
+
+
 }
